+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/customers/CustomerDbConnector.groovy
@@ -7,6 +7,7 @@ import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.AffiliationCategory
 import life.qbic.datamodel.dtos.business.AffiliationCategoryFactory
 import life.qbic.datamodel.dtos.business.Customer
+import life.qbic.datamodel.dtos.general.Person
 import life.qbic.portal.portlet.customers.affiliation.create.CreateAffiliationDataSource
 import life.qbic.portal.portlet.customers.affiliation.list.ListAffiliationsDataSource
 import life.qbic.portal.portlet.customers.create.CreateCustomerDataSource
@@ -249,7 +250,7 @@ class CustomerDbConnector implements CreateCustomerDataSource, UpdateCustomerDat
             "VALUES(?, ?)"
 
     affiliations.each {affiliation ->
-      def affiliationId = getAffiliationId(connection, affiliation)
+      def affiliationId = getAffiliationId(affiliation)
       def statement = connection.prepareStatement(query)
       statement.setInt(1, customerId)
       statement.setInt(2, affiliationId)
@@ -257,9 +258,8 @@ class CustomerDbConnector implements CreateCustomerDataSource, UpdateCustomerDat
 
     }
   }
-
-  //Fixme no use of closeable
-  private static int getAffiliationId(Connection connection, Affiliation affiliation) {
+  
+  int getAffiliationId(Affiliation affiliation) {
     String query = "SELECT * FROM affiliation WHERE organization=? " +
             "AND address_addition=? " +
             "AND street=? " +
@@ -267,18 +267,21 @@ class CustomerDbConnector implements CreateCustomerDataSource, UpdateCustomerDat
             "AND city=?"
 
     List<Integer> affiliationIds = []
+    Connection connection = connectionProvider.connect()
 
-    def statement = connection.prepareStatement(query)
-    statement.setString(1, affiliation.organisation)
-    statement.setString(2, affiliation.addressAddition)
-    statement.setString(3, affiliation.street)
-    statement.setString(4, affiliation.postalCode)
-    statement.setString(5, affiliation.city)
-    statement.execute()
-    ResultSet rs = statement.getResultSet()
-    while (rs.next()) {
-      affiliationIds.add(rs.getInt(1))
+    connection.withCloseable {
+      def statement = connection.prepareStatement(query)
+      statement.setString(1, affiliation.organisation)
+      statement.setString(2, affiliation.addressAddition)
+      statement.setString(3, affiliation.street)
+      statement.setString(4, affiliation.postalCode)
+      statement.setString(5, affiliation.city)
+      ResultSet rs = statement.executeQuery()
+      while (rs.next()) {
+        affiliationIds.add(rs.getInt(1))
+      }
     }
+
 
     if(affiliationIds.size() > 1) {
       throw new DatabaseQueryException("More than one entry found for $affiliation.")
@@ -435,5 +438,31 @@ class CustomerDbConnector implements CreateCustomerDataSource, UpdateCustomerDat
     }
 
     return generatedKeys[0]
+  }
+
+  /**
+   * Searches for a person and returns the matching ID from the person table
+   *
+   * @param person The database entry is searched for a person
+   * @return the ID of the database entry matching the person
+   */
+  int getPersonId(Person person) {
+    String query = "SELECT id FROM person WHERE first_name = ? AND last_name = ? AND email = ?"
+    Connection connection = connectionProvider.connect()
+
+    int personId = -1
+
+    connection.withCloseable {
+      def statement = connection.prepareStatement(query)
+      statement.setString(1, person.firstName)
+      statement.setString(2, person.lastName)
+      statement.setString(3, person.emailAddress)
+
+      ResultSet result = statement.executeQuery()
+      while (result.next()){
+        personId = result.getInt(1)
+      }
+    }
+    return personId
   }
 }