+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/CreateCustomerView.groovy
@@ -9,6 +9,7 @@ import com.vaadin.icons.VaadinIcons
 import com.vaadin.server.UserError
 import com.vaadin.shared.ui.ContentMode
 import com.vaadin.ui.*
+import com.vaadin.ui.themes.ValoTheme
 import groovy.util.logging.Log4j2
 import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.portal.qoffer2.web.controllers.CreateCustomerController
@@ -40,7 +41,8 @@ class CreateCustomerView extends VerticalLayout {
     ComboBox<Affiliation> affiliationComboBox
     ComboBox<Affiliation> addressAdditionComboBox
     Button submitButton
-    Button affiliationButton
+    Button createAffiliationButton
+    Button abortButton
     Panel affiliationDetails
 
     CreateCustomerView(CreateCustomerController controller, ViewModel sharedViewModel, CreateCustomerViewModel createCustomerViewModel) {
@@ -84,12 +86,17 @@ class CreateCustomerView extends VerticalLayout {
         addressAdditionComboBox.setCaption("Address Addition")
         addressAdditionComboBox.enabled = false
 
-        this.affiliationButton = new Button()
-        styleDetailsToggle()
+        this.createAffiliationButton = new Button("Create Affiliation")
+        createAffiliationButton.setIcon(VaadinIcons.INSTITUTION)
 
         this.submitButton = new Button("Create Customer")
-        submitButton.setIcon(VaadinIcons.USER)
+        submitButton.setIcon(VaadinIcons.USER_CHECK)
+        submitButton.addStyleName(ValoTheme.BUTTON_FRIENDLY)
         submitButton.enabled = allValuesValid()
+
+        this.abortButton = new Button("Abort Customer Creation")
+        abortButton.setIcon(VaadinIcons.CLOSE_CIRCLE)
+        abortButton.addStyleName(ValoTheme.BUTTON_DANGER)
 
         this.affiliationDetails = new Panel("Affiliation Details")
 
@@ -107,10 +114,16 @@ class CreateCustomerView extends VerticalLayout {
         row3.setComponentAlignment(addressAdditionComboBox, Alignment.TOP_LEFT)
         row3.setSizeFull()
 
-        HorizontalLayout row4 = new HorizontalLayout(affiliationDetails, submitButton)
-        row4.setComponentAlignment(affiliationDetails, Alignment.BOTTOM_LEFT)
-        row4.setComponentAlignment(submitButton, Alignment.BOTTOM_RIGHT)
+        VerticalLayout affiliationPanel = new VerticalLayout(affiliationDetails)
+        affiliationPanel.setMargin(false)
+        affiliationPanel.setComponentAlignment(affiliationDetails, Alignment.TOP_LEFT)
+        HorizontalLayout buttonLayout = new HorizontalLayout(createAffiliationButton, abortButton,
+                submitButton)
+        buttonLayout.setMargin(false)
+        HorizontalLayout row4 = new HorizontalLayout(affiliationPanel, buttonLayout)
+        row4.setComponentAlignment(buttonLayout, Alignment.BOTTOM_RIGHT)
         row4.setSizeFull()
+
 
         //Add the components to the FormLayout
         this.addComponents(row1, row2, row3, row4)
@@ -228,15 +241,7 @@ class CreateCustomerView extends VerticalLayout {
             submitButton.enabled = allValuesValid()
             addressAdditionComboBox.enabled = Objects.isNull(createCustomerViewModel.affiliation)
         })
-        sharedViewModel.addPropertyChangeListener({it ->
-            switch (it.propertyName) {
-                case "createAffiliationVisible":
-                    styleDetailsToggle()
-                    break
-                default:
-                    break
-            }
-        })
+
         /* refresh affiliation list and set added item as selected item. This is needed to keep this
         field up to date and select an affiliation after it was created */
         sharedViewModel.affiliations.addPropertyChangeListener({
@@ -361,10 +366,6 @@ class CreateCustomerView extends VerticalLayout {
             }
         })
 
-        this.affiliationButton.addClickListener({
-            sharedViewModel.createAffiliationVisible = !sharedViewModel.createAffiliationVisible
-        })
-
         this.affiliationComboBox.addSelectionListener({
             fireAffiliationSelectionEvent(it.value)
             updateAffiliationDetails(it.value)
@@ -383,11 +384,6 @@ class CreateCustomerView extends VerticalLayout {
         content.setMargin(true)
         content.setSpacing(false)
         this.affiliationDetails.setContent(content)
-    }
-
-    private void styleDetailsToggle() {
-        affiliationButton.icon = sharedViewModel.createAffiliationVisible ? VaadinIcons.MINUS : VaadinIcons.PLUS
-        affiliationButton.caption = sharedViewModel.createAffiliationVisible ? "Hide Affiliation Details" : "Show Affiliation Details"
     }
 
     /**