+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/CreateOfferView.groovy
@@ -1,11 +1,10 @@
 package life.qbic.portal.qoffer2.web.views
 
+import com.vaadin.ui.Component
 import com.vaadin.ui.FormLayout
-import com.vaadin.ui.VerticalLayout
 import life.qbic.datamodel.dtos.business.ProductItem
 import life.qbic.portal.qoffer2.web.controllers.CreateOfferController
 import life.qbic.portal.qoffer2.web.controllers.ListProductsController
-
 import life.qbic.portal.qoffer2.web.viewmodel.CreateOfferViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.ProductItemViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.ViewModel
@@ -30,6 +29,7 @@ class CreateOfferView extends FormLayout{
 
     final private ViewModel sharedViewModel
     final private CreateOfferViewModel view
+
     final private CreateOfferController controller
     final private ListProductsController listProductsController
 
@@ -39,73 +39,129 @@ class CreateOfferView extends FormLayout{
     final private SelectItemsView selectItemsView
     final private OfferOverviewView overviewView
 
+    final private CreateCustomerView createCustomerView
+    private ButtonNavigationView navigationView
+    final private CreateAffiliationView createAffiliationView
 
-    CreateOfferView(ViewModel sharedViewModel, CreateOfferViewModel createOfferViewModel, CreateOfferController controller, ListProductsController listProductsController) {
+    final private ViewHistory viewHistory
+
+
+    CreateOfferView(ViewModel sharedViewModel, CreateOfferViewModel createOfferViewModel, CreateOfferController controller, ListProductsController listProductsController, CreateCustomerView createCustomerView, CreateAffiliationView createAffiliationView) {
         super()
         this.sharedViewModel = sharedViewModel
         this.view = createOfferViewModel
         this.controller = controller
         this.listProductsController = listProductsController
+        this.createCustomerView = createCustomerView
+        this.createAffiliationView = createAffiliationView
+        this.projectInformationView = new ProjectInformationView(view)
+        this.customerSelectionView = new CustomerSelectionView(view)
 
-        projectInformationView = new ProjectInformationView(view)
-        customerSelectionView = new CustomerSelectionView(view)
-        projectManagerSelectionView = new ProjectManagerSelectionView(view)
-        selectItemsView = new SelectItemsView(view,sharedViewModel)
-        overviewView = new OfferOverviewView(view)
+        this.projectManagerSelectionView = new ProjectManagerSelectionView(view)
+        this.selectItemsView = new SelectItemsView(view,sharedViewModel)
+        this.overviewView = new OfferOverviewView(view)
 
         initLayout()
         registerListeners()
+        fetchData()
+
+        // add all step views to the main view
+        addStepViewsToCurrent()
+        // hide all views
+        hideViews()
+
+        // Init the view navigation history to be able to navigate back in history
+        this.viewHistory = new ViewHistory(projectInformationView)
     }
 
     /**
      * Initializes the view with the ProjectInformationView, which is the first component to be shown
      */
     private void initLayout(){
+        navigationView = new ButtonNavigationView()
+                .addNavigationItem("1. Project Information")
+                .addNavigationItem("2. Select Customer")
+                .addNavigationItem("3. Assign Project Manager")
+                .addNavigationItem("4. Add Product Items")
+                .addNavigationItem("5. Offer Overview")
+
+        navigationView.showNextStep()
+        this.addComponent(navigationView)
         this.addComponent(projectInformationView)
+        this.setSizeFull()
+    }
+
+    /**
+     * Fetch data from database for observable lists
+     */
+    private void fetchData(){
+        listProductsController.listProducts()
+        //todo add the initalization of data e.g. customer and PM here
     }
 
     /**
      * Registers all listeners for the buttons that enable switching between the different subviews of CreateOfferView and saving the offer
      */
     private void registerListeners() {
+
         this.projectInformationView.next.addClickListener({ event ->
-            this.removeComponent(projectInformationView)
-            this.addComponent(customerSelectionView)
+            viewHistory.loadNewView(customerSelectionView)
+            navigationView.showNextStep()
         })
         this.customerSelectionView.next.addClickListener({
-            this.removeComponent(customerSelectionView)
-            this.addComponent(projectManagerSelectionView)
+            viewHistory.loadNewView(projectManagerSelectionView)
+            navigationView.showNextStep()
         })
         this.customerSelectionView.previous.addClickListener({
-            this.removeComponent(customerSelectionView)
-            this.addComponent(projectInformationView)
+            viewHistory.showPrevious()
+            navigationView.showPreviousStep()
+        })
+        this.customerSelectionView.createCustomerButton.addClickListener({
+            viewHistory.loadNewView(createCustomerView)
+        })
+        this.createCustomerView.abortButton.addClickListener({
+            viewHistory.showPrevious()
+        })
+        this.createCustomerView.submitButton.addClickListener({
+            viewHistory.showPrevious()
+        })
+        this.customerSelectionView.createAffiliationButton.addClickListener({
+            viewHistory.loadNewView(createAffiliationView)
+        })
+        this.createAffiliationView.abortButton.addClickListener({
+            viewHistory.showPrevious()
+        })
+        this.createAffiliationView.submitButton.addClickListener({
+            viewHistory.showPrevious()
         })
         this.projectManagerSelectionView.next.addClickListener({
-            this.removeComponent(projectManagerSelectionView)
-            listProductsController.listProducts()
-            this.addComponent(selectItemsView)
+            viewHistory.loadNewView(selectItemsView)
+            navigationView.showNextStep()
         })
         this.projectManagerSelectionView.previous.addClickListener({
-            this.removeComponent(projectManagerSelectionView)
-            this.addComponent(customerSelectionView)
+            viewHistory.loadNewView(customerSelectionView)
+            navigationView.showPreviousStep()
         })
         this.selectItemsView.next.addClickListener({
-            this.removeComponent(selectItemsView)
-            controller.calculatePriceForItems(getProductItems(view.productItems),view.customerAffiliation.category)
+            controller.calculatePriceForItems(getProductItems(view.productItems),view.customerAffiliation)
             overviewView.fillPanel()
-            this.addComponent(overviewView)
+            viewHistory.loadNewView(overviewView)
+            navigationView.showNextStep()
         })
         this.selectItemsView.previous.addClickListener({
-            this.removeComponent(selectItemsView)
-            this.addComponent(projectManagerSelectionView)
+            viewHistory.loadNewView(projectManagerSelectionView)
+            navigationView.showPreviousStep()
         })
         this.overviewView.previous.addClickListener({
-            this.removeComponent(overviewView)
-            this.addComponent(selectItemsView)
+            viewHistory.loadNewView(selectItemsView)
+            navigationView.showPreviousStep()
         })
         this.overviewView.save.addClickListener({
             controller.createOffer(view.projectTitle, view.projectDescription,view.customer,view.projectManager,
                     getProductItems(view.productItems),view.offerPrice,view.customerAffiliation)
+        })
+        this.createCustomerView.createAffiliationButton.addClickListener({
+            viewHistory.loadNewView(createAffiliationView)
         })
     }
 
@@ -120,5 +176,78 @@ class CreateOfferView extends FormLayout{
             productItems.add(new ProductItem(it.quantity,it.product))
         }
         return productItems
+    }
+
+    private void hideViews() {
+        this.projectInformationView.setVisible(false)
+        this.customerSelectionView.setVisible(false)
+        this.createCustomerView.setVisible(false)
+        this.createAffiliationView.setVisible(false)
+        this.projectManagerSelectionView.setVisible(false)
+        this.selectItemsView.setVisible(false)
+        this.overviewView.setVisible(false)
+    }
+
+    private void addStepViewsToCurrent() {
+        this.addComponents(
+                this.projectInformationView,
+                this.customerSelectionView,
+                this.createCustomerView,
+                this.createAffiliationView,
+                this.projectManagerSelectionView,
+                this.selectItemsView,
+                this.overviewView)
+    }
+
+    /*
+     * Small helper class that assists us keeping track of the view components
+     * history.
+     */
+    private class ViewHistory {
+
+        private List<Component> history
+
+        private int currentPosition
+
+        private Component currentView
+
+        ViewHistory(Component c) {
+            history = new LinkedList<>()
+            currentPosition = 0
+            currentView = c
+            history.add(c)
+            currentView.setVisible(true)
+        }
+
+        def loadNewView(Component view) {
+            history = history.size() > 1 ? history.subList(0, currentPosition+1) : history
+            history.add(view)
+            currentView.setVisible(false)
+            currentView = view
+            currentView.setVisible(true)
+            currentPosition++
+        }
+
+        def showNext() {
+            if(currentPosition == history.size()-1) {
+                // the current view is the latest view in history
+            } else {
+                currentPosition++
+                currentView.setVisible(false)
+                currentView = history.get(currentPosition)
+                currentView.setVisible(true)
+            }
+        }
+
+        def showPrevious() {
+            if(currentPosition == 0) {
+                // the current view is the oldest view in history
+            } else {
+                currentPosition -= 1
+                currentView.setVisible(false)
+                currentView = history.get(currentPosition)
+                currentView.setVisible(true)
+            }
+        }
     }
 }