+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/create/offer/OfferOverviewView.groovy
@@ -7,12 +7,15 @@ import com.vaadin.ui.Grid
 import com.vaadin.ui.HorizontalLayout
 import com.vaadin.ui.Label
 import com.vaadin.ui.Panel
-import com.vaadin.ui.TextArea
-import com.vaadin.ui.TextField
 import com.vaadin.ui.VerticalLayout
+import com.vaadin.ui.renderers.NumberRenderer
 import com.vaadin.ui.themes.ValoTheme
+import life.qbic.datamodel.dtos.business.ProductItem
+import life.qbic.portal.portlet.offers.Currency
 import life.qbic.portal.qoffer2.web.viewmodel.CreateOfferViewModel
 import life.qbic.portal.qoffer2.web.viewmodel.ProductItemViewModel
+
+import java.text.DecimalFormat
 
 /**
  * This class generates a Layout in which the user
@@ -44,8 +47,6 @@ class OfferOverviewView extends VerticalLayout{
      * Initializes the start layout for this view
      */
     private void initLayout(){
-        Label titleLabel = new Label("Offer Overview")
-
         this.previous = new Button(VaadinIcons.CHEVRON_CIRCLE_LEFT)
         previous.addStyleName(ValoTheme.LABEL_LARGE)
 
@@ -64,7 +65,7 @@ class OfferOverviewView extends VerticalLayout{
         this.itemGrid.setItems(createOfferViewModel.productItems)
         generateProductGrid(itemGrid)
 
-        this.addComponents(titleLabel,offerOverview,buttonLayout)
+        this.addComponents(offerOverview,buttonLayout)
     }
 
     /**
@@ -74,11 +75,13 @@ class OfferOverviewView extends VerticalLayout{
      */
     private static void generateProductGrid(Grid<ProductItemViewModel> grid) {
         try {
+
             grid.addColumn({ productItem -> productItem.quantity }).setCaption("Quantity")
             grid.addColumn({ productItem -> productItem.product.productName }).setCaption("Product Name")
             grid.addColumn({ productItem -> productItem.product.description }).setCaption("Product Description")
-            grid.addColumn({ productItem -> productItem.product.unitPrice }).setCaption("Product Unit Price")
-            grid.addColumn({ productItem -> productItem.product.unit.value }).setCaption("Product Unit")
+            grid.addColumn({ productItem -> productItem.product.unitPrice }, new NumberRenderer(Currency.currencyFormat)).setCaption("Product Unit Price")
+            grid.addColumn({ productItem -> productItem.product.unit }).setCaption("Product Unit")
+
 
             //specify size of grid and layout
             grid.setSizeFull()
@@ -92,20 +95,76 @@ class OfferOverviewView extends VerticalLayout{
      * Fills the information of the offer into the panel
      */
     void fillPanel(){
+        /*
+        Container for the complete overview
+         */
         VerticalLayout content = new VerticalLayout()
-        content.addComponent(new Label("${createOfferViewModel.projectTitle}"))
-        content.addComponent(new Label("${createOfferViewModel.projectDescription}"))
-        content.addComponent(new Label("${createOfferViewModel.customer}"))
-        content.addComponent(new Label("${createOfferViewModel.customerAffiliation}"))
-        content.addComponent(new Label("${createOfferViewModel.projectManager}"))
+        /*
+        A header that contains basic project info and the price overview
+         */
+        HorizontalLayout header = new HorizontalLayout()
+        /*
+        The detailed project information container
+         */
+        VerticalLayout projectInfo = new VerticalLayout()
+        projectInfo.addComponent(new Label("${createOfferViewModel.projectTitle}"))
+        projectInfo.addComponent(new Label("${createOfferViewModel.projectDescription}"))
+        projectInfo.addComponent(new Label("${createOfferViewModel.customer}"))
+        projectInfo.addComponent(new Label("${createOfferViewModel.customerAffiliation}"))
+        projectInfo.addComponent(new Label("${createOfferViewModel.projectManager}"))
+        /*
+        Here we set the header components, which is the project info
+        on the left and a basic cost overview on the right
+         */
+        header.addComponent(projectInfo)
+        header.addComponent(createCostOverview())
+        header.setWidthFull()
+        header.setDefaultComponentAlignment(Alignment.TOP_LEFT)
+        /*
+        We add the header as top component in the final view
+         */
+        content.addComponent(header)
         content.addComponent(itemGrid)
 
-        Label price = new Label("total price: ${createOfferViewModel.offerPrice}")
-        content.addComponent(price)
-
-        content.setComponentAlignment(price,Alignment.MIDDLE_RIGHT)
-
         offerOverview.setContent(content)
+    }
+
+    private Panel createCostOverview() {
+        final Panel panel = new Panel("Cost overview")
+        panel.setSizeUndefined()
+        Grid<PriceField> gridLayout = new Grid<>()
+        gridLayout.setHeightByRows(4)
+        gridLayout.setItems([
+                new PriceField("Net price", createOfferViewModel.netPrice),
+                new PriceField("Overheads", createOfferViewModel.overheads),
+                new PriceField("Taxes", createOfferViewModel.taxes),
+                new PriceField("Total price", createOfferViewModel.totalPrice)
+        ])
+        gridLayout.addColumn(PriceField::getName)
+        gridLayout.addColumn(  {
+            costs -> costs.value},
+                new NumberRenderer(Currency.currencyFormat))
+
+        gridLayout.headerVisible = false
+        panel.setContent(gridLayout)
+
+        return panel
+    }
+
+    /*
+    Small helper object, that will display information
+    about individual price positions for offer overviews.
+     */
+    private class PriceField {
+
+        String name
+
+        Double value
+
+        PriceField(String name, Double value) {
+            this.name = name
+            this.value = value
+        }
     }
 
 }