+++ qoffer-infrastructure/src/test/groovy/life/qbic/portal/qoffer2/products/ProductsDbConnectorSpec.groovy
@@ -1,9 +1,13 @@
 package life.qbic.portal.qoffer2.products
 
 import groovy.sql.GroovyRowResult
+import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.services.AtomicProduct
+import life.qbic.datamodel.dtos.business.services.PrimaryAnalysis
 import life.qbic.datamodel.dtos.business.services.Product
+import life.qbic.datamodel.dtos.business.services.ProductUnit
 import life.qbic.portal.portlet.exceptions.DatabaseQueryException
+import life.qbic.portal.qoffer2.customers.CustomerDbConnector
 import life.qbic.portal.qoffer2.database.ConnectionProvider
 import org.apache.groovy.sql.extensions.SqlExtensions
 import spock.lang.Specification
@@ -78,5 +82,44 @@ class ProductsDbConnectorSpec extends Specification {
 
   }
 
+  def "Returns correct id for a given product"() {
+    given: "some expected query results"
+    String query = "SELECT id FROM product "+
+            "WHERE category = ? AND description = ? AND productName = ? AND unitPrice = ? AND unit = ?"
 
+    and: "a connection returning the id of the product if it was found"
+    GroovyMock(SqlExtensions, global: true)
+    SqlExtensions.toRowResult(_ as ResultSet) >> new GroovyRowResult(["id":id])
+
+    PreparedStatement preparedStatement = Mock (PreparedStatement, {
+      it.setString(1, category) >> _
+      it.setString(2, description) >> _
+      it.setString(3, productName) >> _
+      it.setDouble(4, unitPrice) >> _
+      it.setString(5, unit.value) >> _
+
+      it.executeQuery() >> Stub(ResultSet,{it.next() >>> [true, false]})
+    })
+    // the connection must only provide precompiled statements for the expected query template
+    Connection connection = Stub( Connection, {
+      it.prepareStatement(query) >> preparedStatement
+    })
+
+    //and: "a ConnectionProvider providing the stubbed connection"
+    ConnectionProvider connectionProvider = Stub (ConnectionProvider, {it.connect() >> connection})
+
+    //and: "an implementation of the SearchCustomerDataSource with this connection provider"
+    ProductsDbConnector dataSource = new ProductsDbConnector(connectionProvider)
+
+    when:
+    int resultId = dataSource.findProductId(new PrimaryAnalysis(productName,description,unitPrice, unit))
+
+    then:
+    resultId == id
+
+    where: "available products information is as follows"
+    id | category | description | productName | unitPrice | unit
+    0 | "Primary Bioinformatics" | "Sample QC with report" | "Sample QC" | 49.99 | ProductUnit.PER_SAMPLE
+
+  }
 }