+++ CHANGELOG.rst
@@ -9,7 +9,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 ----------------------------------------------
 
 **Added**
-
+* Add the option to create an affiliation while creating an offer
+* Add the option to create a customer while creating an offer
 * Implement SearchCustomer use case
 * Connect CreateAffiliation and CreateCustomer use case
 * Implement GUI for CreateAffiliation use case
@@ -22,6 +23,7 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Created the project using QUBE
 * Implement queries for affiliation creation in database
 * Implement use case _List Packages_
+* Dynamic cost overview upon offer creation
 
 **Fixed**